% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianNetwork-package.R
\docType{package}
\name{BayesianNetwork-package}
\alias{BayesianNetwork-package}
\title{BayesianNetwork: Bayesian Network Modeling and Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A "Shiny"" web application for creating interactive Bayesian Network models, learning the structure and parameters of Bayesian networks, and utilities for classic network analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paulgovan/bayesiannetwork}
  \item \url{http://paulgovan.github.io/BayesianNetwork/}
  \item \url{https://github.com/paulgovan/BayesianNetwork}
  \item Report bugs at \url{https://github.com/paulgovan/BayesianNetwork/issues}
}

}
\author{
\strong{Maintainer}: Paul Govan \email{paul.govan2@gmail.com} (\href{https://orcid.org/0000-0002-1821-8492}{ORCID}) [copyright holder]

}
\keyword{internal}
