% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianNetwork.R
\name{BayesianNetwork}
\alias{BayesianNetwork}
\title{Bayesian Network modeling and analysis.}
\usage{
BayesianNetwork()
}
\description{
BayesianNetwork is a Shiny web application for Bayesian Network modeling and
analysis.
}

