% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianNetwork.R
\name{BayesianNetwork}
\alias{BayesianNetwork}
\title{Bayesian Network modeling and analysis.}
\usage{
BayesianNetwork()
}
\value{
The function returns no value. It launches a shiny web application for
Bayesian Network modeling and analysis.
}
\description{
BayesianNetwork is a Shiny web application for Bayesian Network modeling and
analysis.
}
\examples{
if (interactive()) {
  BayesianNetwork()
}
}
\seealso{
\url{http://paulgovan.github.io/BayesianNetwork/}
}
